extract into getYearFilters
