+++ packages/components/src/templates/next/layouts/Collection/filterUtils/getYearFilters.ts
@@ -1,0 +1,41 @@
+import type { FilterItem } from "../../../types/Filter"
+import { FILTER_ID_YEAR, NO_SPECIFIED_YEAR_FILTER_ID } from "./constants"
+
+export const getYearFilters = ({
+  years,
+  numberOfUndefinedDates,
+}: {
+  years: Record<string, number>
+  numberOfUndefinedDates: number
+}): {
+  id: string
+  label: string
+  items: FilterItem[]
+} => {
+  const yearFilterItems = Object.entries(years)
+    .map(([label, count]) => ({
+      id: label.toLowerCase(),
+      label,
+      count,
+    }))
+    .sort((a, b) => parseInt(b.label) - parseInt(a.label))
+
+  return {
+    id: FILTER_ID_YEAR,
+    label: "Year",
+    // do not show "not specified" option if all items have undefined dates
+    items:
+      yearFilterItems.length === 0
+        ? []
+        : numberOfUndefinedDates === 0
+          ? yearFilterItems
+          : [
+              ...yearFilterItems,
+              {
+                id: NO_SPECIFIED_YEAR_FILTER_ID,
+                label: "Not specified",
+                count: numberOfUndefinedDates,
+              },
+            ],
+  }
+}
